optimised transactions between menus using stack. made a transaction between menu and game.
