+++ Assets/Scripts/MoveMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4e9ee7f8d03a214a8cbe92248cf80f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
