+++ Assets/Scripts/UIDissolve.cs
@@ -3,9 +3,16 @@ using System.Collections.Generic;
 using UnityEngine;
 using DG.Tweening;
 using UnityEngine.UI;
+
 public class UIDissolve : MonoBehaviour
 {
+
+                    // too dirty script
+                    //must be rewritten!
+
+
     [SerializeField] private GameObject[] UI;
+    GameObject canvas;
     [SerializeField] private float travelTime;
     [SerializeField] private float travelDistance;
     [SerializeField] private float knifeTravelTime;
@@ -13,30 +20,38 @@ public class UIDissolve : MonoBehaviour
 
     private void Start()
     {
+        canvas = transform.parent.gameObject;
+        canvas.SetActive(true);
         MoveOut();
     }
     private void MoveOut()
     {
-        foreach(GameObject UIElement in UI)
+        foreach (GameObject UIElement in UI)    //dependency from literals must be disposed. Even looks strange
         {
             switch (UIElement.name)
             {
-                case "StartButton" : Destroy(UIElement);
+                case "StartButton":
+                    Destroy(UIElement);
                     break;
 
                 case "Knife":
-                    UIElement.transform.DOMoveY(UIElement.transform.position.y + knifeTravelDistance*-1, knifeTravelTime);
+                    UIElement.transform.DOMoveY(UIElement.transform.position.y + knifeTravelDistance * -1, knifeTravelTime);
                     break;
 
                 default:
-                    UIElement.transform.DOMoveX(UIElement.transform.position.x + travelDistance, travelTime, false);  
+                    UIElement.transform.DOMoveX(UIElement.transform.position.x + travelDistance, travelTime, false);
                     UIElement.GetComponent<Image>().CrossFadeAlpha(0, travelTime, false);
-
-                    travelDistance *= -1;
                     break;
             }
-            
+            travelDistance *= -1;
         }
-    }    
+        StartCoroutine(WaitForUI(travelTime));
+        
+    }
+    private IEnumerator WaitForUI(float time)
+    {
+        yield return new WaitForSeconds(time);
+        canvas.SetActive(false);
+    }
 }
